# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: los-ast API
  description: |
    los-ast 服务化 API - 代码智能分析内核的 HTTP façade

    ## 硬约束清单
    1. **P0 Scope Freeze**: Core 稳定能力固定为 `/healthz/*`, `/scan`, `/discover/symbols`
    2. **Core/API 硬边界**: Core 只返回领域错误，API 负责协议转换
    3. **Scope 验证**: 生产环境强制完整 scope，拒绝 `mode: local`
    4. **扫描限制**: 最大 1000 文件 / 10MB 响应 / 30秒超时
    5. **取消语义**: 区分 client disconnect vs server timeout

  version: 1.0.0
  contact:
    name: los-ast Team
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
...
  /vps-agent-web/approvals:
    get:
      summary: 查询审批项
      description: |
        VPS Agent Web 对接前缀下的审批查询接口。
        请求会按 scope 做租户隔离。
      tags:
        - VPS Agent Web
      x-stability: beta
      x-migration-target: VPS Agent Web
      parameters:
        - $ref: '#/components/parameters/ScopeQuery'
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, expired]
        - name: item_type
          in: query
          schema:
            type: string
            enum: [recovery_action, code_patch, config_change, recipe_activation]
        - name: risk_level
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalListResponse'
        '400':
          description: 参数错误或缺少 scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: 创建审批项
      tags:
        - VPS Agent Web
      x-stability: beta
      x-migration-target: VPS Agent Web
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApprovalRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalItemEnvelope'
        '400':
          description: 参数错误或 scope 不完整
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /vps-agent-web/approvals/{id}:
    get:
      summary: 获取审批项详情
      tags:
        - VPS Agent Web
      x-stability: beta
      x-migration-target: VPS Agent Web
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/ScopeQuery'
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalItemEnvelope'
        '404':
          description: 审批项不存在或无权限访问
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /vps-agent-web/approvals/{id}/process:
    post:
      summary: 处理审批项
      tags:
        - VPS Agent Web
      x-stability: beta
      x-migration-target: VPS Agent Web
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/ScopeQuery'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcessApprovalRequest'
      responses:
        '200':
          description: 处理成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalItemEnvelope'
        '400':
          description: 处理失败或参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /vps-agent-web/approvals/stats:
    get:
      summary: 获取审批统计
      tags:
        - VPS Agent Web
      x-stability: beta
      x-migration-target: VPS Agent Web
      parameters:
        - $ref: '#/components/parameters/ScopeQuery'
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalStatsResponse'
        '400':
          description: 缺少 scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    Scope:
      type: object
      description: |
        请求范围标识。生产环境需要完整 scope，开发环境允许 local 模式。
      properties:
        tenant_id:
          type: string
          description: 租户 ID（生产环境必填）
        project_id:
          type: string
          description: 项目 ID（生产环境必填）
        actor_id:
          type: string
          description: 执行者 ID（生产环境必填）
        mode:
          type: string
          enum: [local, service]
          description: |
            执行模式。`local` 仅在开发环境允许，生产环境会被拒绝。
      required:
        - tenant_id
        - project_id
        - actor_id
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - category
            - code
            - message
            - requestId
            - timestamp
            - retryable
          properties:
            category:
              type: string
              enum: [VALIDATION, SCOPE, TIMEOUT, SCAN_TOO_LARGE, NOT_FOUND, SERVICE_UNAVAILABLE, INTERNAL]
            code:
              type: string
            message:
              type: string
            requestId:
              type: string
            timestamp:
              type: string
              format: date-time
            retryable:
              type: boolean
            details:
              type: object
              additionalProperties: true
    CreateApprovalRequest:
      type: object
      required:
        - item_type
        - item_id
        - title
        - description
        - risk_level
        - timeout_seconds
        - scope
      properties:
        item_type:
          type: string
          enum: [recovery_action, code_patch, config_change, recipe_activation]
        item_id:
          type: string
        title:
          type: string
        description:
          type: string
        risk_level:
          type: string
          enum: [low, medium, high, critical]
        timeout_seconds:
          type: integer
          minimum: 1
        scope:
          $ref: '#/components/schemas/Scope'
        metadata:
          type: object
          additionalProperties: true
    ProcessApprovalRequest:
      type: object
      required:
        - action
        - actor_id
      properties:
        action:
          type: string
          enum: [approve, reject]
        actor_id:
          type: string
        comment:
          type: string
    ApprovalItem:
      type: object
      required:
        - approval_id
        - item_type
        - item_id
        - title
        - description
        - risk_level
        - status
        - timeout_at
        - scope
        - created_at
        - updated_at
        - version
      properties:
        approval_id:
          type: string
        item_type:
          type: string
          enum: [recovery_action, code_patch, config_change, recipe_activation]
        item_id:
          type: string
        title:
          type: string
        description:
          type: string
        risk_level:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [pending, approved, rejected, expired]
        requester:
          type: object
          properties:
            actor_id:
              type: string
            timestamp:
              type: string
              format: date-time
        approver:
          type: object
          nullable: true
          properties:
            actor_id:
              type: string
            timestamp:
              type: string
              format: date-time
            comment:
              type: string
        timeout_at:
          type: string
          format: date-time
        scope:
          type: object
          required:
            - tenant_id
            - project_id
          properties:
            tenant_id:
              type: string
            project_id:
              type: string
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        version:
          type: integer
    ApprovalItemEnvelope:
      type: object
      required:
        - approval
      properties:
        approval:
          $ref: '#/components/schemas/ApprovalItem'
    ApprovalListResponse:
      type: object
      required:
        - items
        - total
        - has_more
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ApprovalItem'
        total:
          type: integer
        has_more:
          type: boolean
        next_offset:
          type: integer
    ApprovalStats:
      type: object
      required:
        - total
        - by_status
        - by_risk_level
        - by_type
        - avg_decision_time_seconds
      properties:
        total:
          type: integer
        by_status:
          type: object
          additionalProperties:
            type: integer
        by_risk_level:
          type: object
          additionalProperties:
            type: integer
        by_type:
          type: object
          additionalProperties:
            type: integer
        avg_decision_time_seconds:
          type: number
    ApprovalStatsResponse:
      type: object
      required:
        - stats
      properties:
        stats:
          $ref: '#/components/schemas/ApprovalStats'
...
  parameters:
    ScopeQuery:
      name: scope
      in: query
      required: true
      description: URL 编码 JSON 字符串，需包含 tenant_id/project_id，部分写接口需要 actor_id
      schema:
        type: string
...